backend/requirements.txt
# ── Web framework ──────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1

# ── Classical ML ───────────────────────────────
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2

# ── Deep Learning ──────────────────────────────
transformers==4.41.0
sentence-transformers==3.0.0
torch==2.3.0

# ── Geocoding (setup only) ─────────────────────
geopy==2.4.1
requests==2.32.2
